% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_test_to_dataframe.R
\name{convert_test_to_dataframe}
\alias{convert_test_to_dataframe}
\title{Convert a list of t-test results into a dataframe}
\usage{
convert_test_to_dataframe(results, ts_features, concepts)
}
\arguments{
\item{results}{A list of t-test results}

\item{ts_features}{A vector of features}

\item{concepts}{A vector of concepts}
}
\value{
A dataframe containing the results
}
\description{
Convert a list of t-test results into a dataframe
}
