% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_t_test.R
\name{feat_t_test}
\alias{feat_t_test}
\title{Perform a t-test for a given feature and concept}
\usage{
feat_t_test(df, cluster_var, concept, ts_feature)
}
\arguments{
\item{df}{A data frame with the data}

\item{cluster_var}{The name of the clustering variable in df}

\item{concept}{The name of the current concept}

\item{ts_feature}{The name of the current feature}
}
\value{
A t-test result
}
\description{
Perform a t-test for a given feature and concept
}
