% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_missing.R
\name{feature_missing}
\alias{feature_missing}
\title{Analysis and Visualization of Missing Values in Features}
\usage{
feature_missing(features, title = "Feature-wise Missingness")
}
\arguments{
\item{features}{A data frame containing features with potential missing values.}

\item{title}{A string to be used as the title for the histogram.}
}
\value{
A list with three elements:
   * miss_overall: The overall percentage of missing values in the features data frame.
   * miss_per_feature: A data frame detailing the percentage of missing values per feature.
   * plt_miss_per_feature: A ggplot object representing the histogram of missing value percentages.
}
\description{
This function analyzes the missing values in a given features data frame,
calculates the percentage of missing values, and visualizes these percentages
using a histogram. It returns a list containing the overall percentage of missing
values, a data frame detailing the percentage of missing values per feature,
and the generated plot.
}
\examples{
\dontrun{
data <- data.frame(a = c(1, 2, NA), b = c(4, NA, NA))
missing_values_analysis <- feature_missing(data, "Missing Values Analysis")
}
}
