% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureExtractor.R
\name{featureExtractor}
\alias{featureExtractor}
\title{Feature Extraction Function}
\usage{
featureExtractor(data, items, pid, tid)
}
\arguments{
\item{data}{A data frame from which to extract features.}

\item{items}{A character vector of the names of the variables for which to compute features.}

\item{pid}{A character string specifying the name of the person identifier variable.}

\item{tid}{A character string specifying the name of the time identifier variable.}
}
\value{
A list containing:
   * features: A data frame of raw features.
   * featuresZ: A data frame of standardized features.
   * featuresZMat: A data frame of standardized features with person identifiers as row names.
   * gam: A list of fitted GAM models for each person.
}
\description{
This function extracts features from a given dataset, grouping it by an id variables, and summarizing it into the time series features
for each item in the given list of items. It generates univariate features such as mean, median, standard deviation,
mean absolute deviation, root mean square of successive differences, and mean absolute change.
It also generates multivariate, applying linear and non-linear trends using generalized additive models (GAMs),
and calculates auto-regressive components and their confidence intervals.
It standardizes the features, and outputs a list containing raw and standardized features, as well as the GAM models.
}
\examples{
\dontrun{
featureExtractor(df, c("item1", "item2"), "ID", "time")
}
}
