% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_feature_plots.R
\name{generate_feature_plots}
\alias{generate_feature_plots}
\title{Generate plots based on t-test results}
\usage{
generate_feature_plots(df)
}
\arguments{
\item{df}{An annotated dataframe of t-test results}
}
\value{
A list of ggplots
}
\description{
Generate plots based on t-test results
}
