% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pColMiss.R
\name{pColMiss}
\alias{pColMiss}
\title{Calculate Column Percentage of Missing Values}
\usage{
pColMiss(x)
}
\arguments{
\item{x}{A vector or data frame column in which to calculate the percentage of missing values.}
}
\value{
A numeric value representing the percentage of missing values in the input vector or column.
}
\description{
This function calculates the percentage of missing values in a given vector or data frame column.
}
\examples{
\dontrun{
column <- c(1, 2, NA)
column_missing_percentage <- pColMiss(column)
}
}
