% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_tests.R
\name{run_all_tests}
\alias{run_all_tests}
\title{Run t-tests for all provided features and concepts}
\usage{
run_all_tests(df, cluster_var, ts_features, concepts)
}
\arguments{
\item{df}{A data frame with the data}

\item{cluster_var}{The name of the clustering variable in df}

\item{ts_features}{A vector of features}

\item{concepts}{A vector of concepts}
}
\value{
A list containing t-test results
}
\description{
Run t-tests for all provided features and concepts
}
