% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_Publication.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Create a custom publication colour scale for ggplot}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{...}{Other arguments passed on to layer functions}
}
\value{
A discrete colour scale for ggplot with predefined colour palette.
}
\description{
Create a custom publication colour scale for ggplot
}
\examples{
\dontrun{
library(ggplot2)
ggplot(iris, aes(Species, fill = Species)) + geom_bar() + scale_colour_Publication()
}

}
