% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_Publication.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Create a custom publication fill scale for ggplot}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{...}{Other arguments passed on to layer functions}
}
\value{
A discrete fill scale for ggplot with predefined colour palette.
}
\description{
Create a custom publication fill scale for ggplot
}
\examples{
\dontrun{
library(ggplot2)
ggplot(iris, aes(Species)) + geom_bar(aes(fill = Species)) + scale_fill_Publication()
}
}
