% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Create a custom publication theme for ggplot}
\usage{
theme_Publication(base_size = 14, base_family = "")
}
\arguments{
\item{base_size}{A number to set the base font size. Default is 14.}

\item{base_family}{A character string specifying the base font family. Default is empty.}
}
\value{
A complete ggplot theme with the specified characteristics.
}
\description{
Create a custom publication theme for ggplot
}
\examples{
\dontrun{
library(ggplot2)
ggplot(iris, aes(Sepal.Length, Petal.Length)) + geom_point() + theme_Publication()
}
}
